Refactor literals and puzzle to allow for arbitrary-length literals
This refactors literals to be arbitrary-length. This means we won't
have to make multiple Smoothie classes, for example, and can just
select however many we need at puzzle-generation time.

That required a change to the Puzzle class, which previously had
assumed that the literals used in the puzzle were simply the union
of all the members of the literal types (i.e., all colors, all
drinks, all smoothies, etc.).

This change preserves that as the default assumption, but allows us
to pass a new `elements` kwarg that says "hey, only use these literals
instead of all of them."

This also helps us to programatically generate puzzles of different
sizes, since this effectively removes coupling between the length of
the Literal classes and the number of houses in the puzzle.