lint: add download for Python 3.12, explicitly disable linters
I don't know why it seemed checkov was trying to be installed every run, but in the interest of getting this actually running ... I'll just disable it. And isort / black are unnecessary (ruff supersedes them), bandit is annoying (why are you warning about asserts in test files?and yammllint is not necessary when I don't have yaml. Wait, trunk config is yaml.