refactor (?): try to simplify puzzle initialization
but mostly this is to see what trunk does; it supposedly runs on PRs, so let's try that.

The actual functional effect of this change is that we take a step towards removing the redundancy of the Puzzle initialization; the solution exactly defines the puzzle elements & types of elements.

This creates two unused variables in the puzzle init (`element_types` and `elements`), but that's fine for an intermediate refactor step.